Option to prune stale SLXs on upload to the platform
Specify the --prune-stale-slsx flag when executing the workspace builder to
enable this upload mode. If enabled, then any SLXs in the existing repo
that were originally generated by the workspace builder but are not present
in the new upload data are pruned from the repo. SLXs that were manually
created/configured by the user are never pruned.