Option to prune stale SLXs on upload to the platform (#468)
* Option to prune stale SLXs on upload to the platform

Specify the --prune-stale-slsx flag when executing the workspace builder to
enable this upload mode. If enabled, then any SLXs in the existing repo
that were originally generated by the workspace builder but are not present
in the new upload data are pruned from the repo. SLXs that were manually
created/configured by the user are never pruned.

* add prune-slx option to ui, add rule for cluster resources

* raise error if workspaceName is not string
https://github.com/runwhen-contrib/runwhen-local/issues/470

* Add/kubeconfig-secret-management (#472)

* I hate testing after builds...

* cut the workflow bloat (for now)

* change secret name and switch to yaml

* fix up cacert

* add update logic

* x

* fix secret name

* version bump

---------

Signed-off-by: Shea Stewart <shea.stewart@runwhen.com>
Co-authored-by: Rob Vaterlaus <rob.vaterlaus@runwhen.com>
Co-authored-by: Shea Stewart <stewart.shea@gmail.com>
Co-authored-by: Shea Stewart <shea.stewart@runwhen.com>